chore(PR): extract data fetching to a react hook
